requirements.txt
feedparser
google-generativeai
requests
beautifulsoup4
markdown
